Move epoxy related poll functions back to MessageItemFactory.
